format and rename test examples file
